Don't recursively delete old MSU pack
The shuffler deletes the old shuffled MSU pack before generating a new one, in case the new shuffled pack needs to have fewer tracks than the old one, so stale tracks from the old one aren't left around.

This was done recursively, which is fine if the script is run as intended (in its own subdirectory), but an understandable user error running the script in the same directory containing other MSU packs would result in deleting the entire MSU collection.

Credit to Falcon for finding the bug, and sorry for deleting your whole MSU collection :(