Recursively find packs for extended tracks
The previous version would only search for packs to match for extended MSU tracks in ../*/*.pcm in an attempt to avoid getting confused by packs with subdirectories with optional tracks.  It would also consider the partially completed shuffled MSU pack in progress as an option for a non-extended pack.

This resulted in the shuffler choosing its own generic dungeon/boss music as the dungeon/boss-specific track, and failed to find any packs in deeper directory structures like ../*/*/*.pcm

This change fixes both of these behaviors; thanks ScorelessPine for the bug report.