Fix forcerealcopy bug for Python 3.8 users
The code that checks whether the script is on the same drive as the ROM relies on Python 3.9-exclusive behavior that makes __file__ always return an absolute path.  Because of this, Python 3.8 (or earlier) users would sometimes be forced to use real copies instead of hardlinks in situations that didn't require it.  This change fixes this bug, bumps the version, and changes output to always print the version for easier debugging in the future.